HOCS-2461: change custom reports to stream
Currently on the database we pull all the data and then do the 
data conversion. This leads to a big lead time on the download as 
we currently handle this linearly. 
This change makes it so that we stream the data from the custom 
view and thus start building the file as the data comes through.