HOCS-2461: change initialisation of adapters
Previously we were initialising the adapters on every request, 
the new changes that were added changed this to the constructor 
of the `CustomExportDataConverter`. This is an issue as we need 
to keep the cached information up to date. This change therefore 
changes the location that we update this information so it happens 
on every export request.