HOCS-2624 start streaming file after writing headers (#84)
* HOCS-2624: start streaming after header write

At present we wait for the first set of data to come through 
before we start streaming the custom audit file to the user. 
This change flushes the file to the user straight after creating 
it, bypassing the initial cut off on the file download if the results 
take longer than 2 minutes to return,

* HOCS-2624: lower fetch size 

At present there is still an issue with some results for 
custom reports taking longer than two minutes to 
return. Therefore, lowering the batch size down to 20 
should assist with this.