HOCS-2310: change custom export route (#89)
* HOCS-2310: remove header check

As we are not routing the endpoint through Keycloak anymore and 
the endpoint itself is not external. We don't need to protect this
endpoint.

* HOCS-2310: change routing 

This change changes the routing so that it doesn't go through 
keycloak gatekeeper. This then allows us to use basicauth instead of 
fudging our way around roles and whitelisting the endpoint.

* HOCS-2310: remove unnecessary test

As we are no longer using RBAC'd group security on this endpoint, 
we no longer need to have this test case.