Migrate to Java 11
Migrate the build.gradle file to Java 11. Gradle 7 no longer supports the usage of compile and testCompile. These have been replaced with implementation and test implementation. 
Too allow for lombok to work correctly we now need to specify a 
annotationProcessor for it. 
As we are now targeting Java 11 we set the `sourceCompatibility` 
and the `targetCompatibility` to 11