Lock OpenJDK8 Version
A recent update to the base image of this branch meant that it's 
now being built using CentOS 8 instead of CentOS 7. This is causing 
our docker builds to fail so this change pegs the version to the 
previous version while we identify the changes required for this to 
now work.