Revert "Java11 migration (#91)" (#93)
* Revert "Java11 migration (#91)"

Issues were identified whereby spring boot would not recognise a 
correct route. As a result the ability to pull reports has been 
compromised. As a result we need to revert this migration while 
further research into the potential migration take place. 

This reverts commit c4bb74226a4681f1667b1c2b1b7b8ade0d7ff0ba.

* Lock OpenJDK8 Version

A recent update to the base image of this branch meant that it's 
now being built using CentOS 8 instead of CentOS 7. This is causing 
our docker builds to fail so this change pegs the version to the 
previous version while we identify the changes required for this to 
now work.