HOCS-3120: add max header size property
As part of the spring-boot 2.0.5 to 2.1.4 migration the max header 
size for all 4 of the available embedded servers has been made 
consistent. The new maximum is 8kb. As we route our request 
directly through the keycloak-gatekeeper the headers that also 
adds exceeds this new limit and the service throws a 400. While 
we investigate how big the header size is, this change reverts the 
max header size back to the previous default of 1Mb for undertow.
https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.1-Release-Notes#consistent-max-http-header-size-across-all-embedded-web-servers