HOCS-3120: migrate to Java 11
* Upgrade Gradle Version

Java 11 is currently not support through `gradle-4-5-1-all`. As Java 11
is support from gradle 5 onwards this change brings it up to the
latest released gradle version.

* Migrate to Java 11

Migrate the build.gradle file to Java 11. Gradle 7 no longer supports the usage of compile and testCompile. These have been replaced with implementation and test implementation.
Too allow for lombok to work correctly we now need to specify a
annotationProcessor for it.
As we are now targeting Java 11 we set the `sourceCompatibility`
and the `targetCompatibility` to 11

* Change the docker image to OpenJDK11

Currently the docker image used as a base is `openjdk8`. As we are
now supporting the same Java 11 as the other services, this base
image is using the same as those: `openjdk11:v11.0.5_10`.

* Bump dependencies versions

Currently audit is running behind the other services on the
springBootVersion and the camelVersion. This change brings it more
in line and up to the highest patch version of the minor version.

* Change drone.yaml image version

As we are now building for Java 11 we have to use the
`openjdk11:v11.0.5_10` image in the drone pipeline.