HOCS-3083: make initialise stateless
Currently as we use this class as a Service, it's underlying a 
singleton. In the initialise function which is called on every 
converted request, we clear out the conversion maps before 
re-populating. 
If however another request is using the data from this when 
converting, there is chance that the data is not present and 
therefore defaults to showing the uuid's instead of the converted 
data. This change therefore makes it so that the state of the 
conversion data does not get restored. Instead we now do an 
update to the data for every item.