HOCS-3083: remove caching from info client
The info client currently caches the results of the info service calls. 
As we now allow for changes within the management ui, these 
cached calls mean that some audit results might continue to show 
out of date information. Removing this therefore enforces that all 
the data is in sync across multiple replicas.