HOCS-3083: abstract out initialisation of converter
To keep the single responsibility principle and clean up the code, 
this change moves the responsibility of the initialisation out of the 
ExportDataConverter and into a factory class. This simplifies the 
construction of this and removes the need to have a initialiser in 
every case data call. 
This change also simplifies all the test cases as we now pass maps 
into the ExportDataConverter.