HOCS-3083: UUID showing incorrectly in converted cases (#102)
Currently as we use the ExportDataConverter class as a Service, 
it's underlying a singleton. In the initialise function which is 
called on every converted request, we clear out the 
conversion maps before re-populating. This is causing issues 
whereby if two requests come through at the same time, the 
hashmap that contains the uuid's can be empty and therefore 
the uuid would be used.

To remediate this, the initialisation has been extracted out 
into a factory class that returns a new ExportDataConverter
with the initialised data. This removes the issue of not having 
data at the time of the conversion.

Other improvements include removing code that isn't used, 
removing caching on functions that can have changind data.