HOCS-3948: cleanup properties and create queue
Add the queue creation on the initial startup when using localstack.
This also clears the queue to ensure that tests have a clear slate to
run from incase the queue already exists. Swap from .properties file to
use yaml as clearer to understand.
Rename property names to become more consistent.