HOCS-3948: Camel Queue and Processing Removal (#112)
* HOCS-3948: rewrite build.gradle file

The build.gradle file currently uses an outdated mechanism with the use
of a buildscript block. This change reworks the file.

* HOCS-3948: remove findbugs and checkstyle

Currently, we have two config files that are unused and can be removed.

* HOCS-3948: remove camel dependency and usages

Remove all usages of camel from the service.

* HOCS-3948: add spring cloud aws messaging

Add the spring cloud aws messaging dependency.

* HOCS-3948: amend sqs and LocalStack config files

Ammend configuration files to use async client builders and to supply an
SimpleMessageListenerContainerFactory bean.
Replace deprecated isEmpty with hasText.

* HOCS-3948: clean up Spring Configuration

* HOCS-3948: add GsonConfiguration bean

Add a GsonConfiguration that provides a Gson bean that supplies a
LocalDateTimeDeserializer.

* HOCS-3948: change to use GSON Annotations

As we are now using GSON for the deserializing of creat audit events,
this changes the use of Jackson annotations to GSON ones.

* HOCS-3948: add SqsListener

Add a sqs listener that handles the creating of audit events. This
parses the message provided to a CreateAuditDto and then handles that
through to the AuditDataService.

* HOCS-3948: amend application.properties

Cleanup the application properties to remove SEDA queues, unnecessary
redrive policy and other properties. Rename usages in configuration file.

* HOCS-3948: add tests for audit queues and listener

* HOCS-3948: rework build.gradle and change prop usage

Reorganise build.gradle for consistency and update all dependencies to
the latest version. Flyway cannot be updated due to the community issue
of 8.0.0 not support Postgres 9.6. Appears the Postgres image doesn't
exists either, so this dependency hasn't been updated either.

* HOCS-3948: cleanup properties and create queue

Add the queue creation on the initial startup when using localstack.
This also clears the queue to ensure that tests have a clear slate to
run from incase the queue already exists. Swap from .properties file to
use yaml as clearer to understand.
Rename property names to become more consistent.

* HOCS-3948: update drone.yaml file

Add localstack service to the drone.yaml and cleanup the test step to
pass through the env and not through commands.

* HOCS-3948: update docker-compose file

* HOCS-3948: PR Comments