HOCS-4215: add data parser
When exports are run with the `convert=true` flag a ExportDataConverter
is created. This is then used against every column for every row of an
extract to attempt to convert. This is highly inefficient as we know
(apart from CASE_DATA extracts) what we want to display and the values
that should be converted. Even more so is if its a UUID and not in the
conversion map - as this calls casework as it presumes its a case UUID.

This change adds a data parser (with associated factory) for BF/FOI
Interest extracts that supports the conversion of only the values that
we have to. This allows for each case type to specify which values
should be stored for conversion. Overall this improves the overall
efficiency of the extracts as it only initialises what it needs.