HOCS-5038: improve latest stream handling
There is an issue with the implementation of the streaming with the
`audit_event_latest_events` query whereby upserting information during a
report pull could cause duplicate rows to be returned or missing data.

This change adds row locking when querying this table for the rows that
are returned. This stops any updates to happen while the data is being
read. As queries take less than 1 second on average, this should be more
than enough for most cases.

Improvements to the query have also been added so that inserts after the
report has been started in the database aren't returned - removing
post-added data from causing issues.