HOCS-5113: change to use request date time
There is currently an edge case whereby if a request comes through that
the data in the export converter is not upto date. For example if a
report is being generated but an action on a new case occurs in this
time the case UUID is not present.

This is leading to UUIDs being presented to the user instead of the
references. This change swaps out the way we work out the end date to
use the request `LocalDateTime::now()`, this means any new data during
this report pull won't be included. This along with ensuring we get the
converter after the data stream ensure thats all UUIDs are present.