HOCS-5279: add case data fields configuration (#308)
When calling the CASE_DATA extract the service currently calls out to
hocs-info-service and retrieves all the fields from a custom table and
screens for that case type. This adds a dependency on the
`hocs-info-service` and also is prone to mistakes leading to missing
fields or extra fields.

This change adds a configuration file that contains all the additional
fields that should be appended to the extract. This not only removes one
of the dependencies to `hocs-info-service`, but gives greater control
of fields and there appearance in the extract.