HOCS-5279: add custom export view configuration
When conducting a custom export, there are calls to hocs-info-service to
retrieve some required information to conduct the request. This isn't
sensitive information (as we already have the schema in this repository)
and can be migrated into this repository as a configuration file.

This change implements this and migrates the required configuration into
the repository. All methods around this have been updated to utilise the
new config schema.