HOCS-6146: move materialised refresh to controller
Move the `refreshMaterialisedView` method to its own controller not
imposed by the `extracts` profile. While this works - the imposed 1
minute timeout kills the underlying refresh.

Moving this to its own controller allows for the main `hocs-audit`
deployment to trigger the refresh. This still has the 1 minute timeout
applied - but this can be turned off for now.